requirements.txt
Django==1.6.3
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==18.0
pystache==0.5.3
static==1.0.2
wsgiref==0.1.2
South==0.8.4
django-social-auth==0.7.28
httplib2==0.8
oauth2==1.5.211
oauthlib==0.6.1
python-openid==2.2.5
python-social-auth==0.1.21
requests==2.2.1
requests-oauthlib==0.4.0
selenium==2.39.0
six==1.5.2